+++ ElementX/Sources/Screens/Spaces/SpaceListScreen/SpaceListScreenViewModel.swift
@@ -11,17 +11,26 @@ import SwiftUI
 typealias SpaceListScreenViewModelType = StateStoreViewModelV2<SpaceListScreenViewState, SpaceListScreenViewAction>
 
 class SpaceListScreenViewModel: SpaceListScreenViewModelType, SpaceListScreenViewModelProtocol {
+    private let spaceServiceProxy: SpaceServiceProxyProtocol
+    
     private let actionsSubject: PassthroughSubject<SpaceListScreenViewModelAction, Never> = .init()
     var actionsPublisher: AnyPublisher<SpaceListScreenViewModelAction, Never> {
         actionsSubject.eraseToAnyPublisher()
     }
 
-    init(userSession: UserSessionProtocol) {
+    init(userSession: UserSessionProtocol, spaceServiceProxy: SpaceServiceProxyProtocol) {
+        self.spaceServiceProxy = spaceServiceProxy
+        
         super.init(initialViewState: SpaceListScreenViewState(userID: userSession.clientProxy.userID,
-                                                              rooms: [],
+                                                              joinedSpaces: spaceServiceProxy.joinedSpacesPublisher.value,
                                                               joinedRoomsCount: 0,
                                                               bindings: .init()),
                    mediaProvider: userSession.mediaProvider)
+        
+        spaceServiceProxy.joinedSpacesPublisher
+            .receive(on: DispatchQueue.main)
+            .weakAssign(to: \.state.joinedSpaces, on: self)
+            .store(in: &cancellables)
         
         userSession.clientProxy.userAvatarURLPublisher
             .receive(on: DispatchQueue.main)
@@ -40,6 +49,10 @@ class SpaceListScreenViewModel: SpaceListScreenViewModelType, SpaceListScreenVie
         MXLog.info("View model: received view action: \(viewAction)")
         
         switch viewAction {
+        case .spaceAction(.select(let spaceRoom)):
+            actionsSubject.send(.selectSpace(spaceRoom))
+        case .spaceAction(.join(let spaceRoom)):
+            #warning("Implement joining.")
         case .showSettings:
             actionsSubject.send(.showSettings)
         }